refactor: rename `TimeSeries.ts` to `time-series.model.ts`
