+++ back-end-challenge-v2-solution/src/models/time-series.model.ts
